requirements.txt
Flask
torch
torchvision
timm
Pillow
opencv-python
Werkzeug
